pyproject.toml
[project]
name = "inventorysystem-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.72.1",
    "fastapi[standard]>=0.121.2",
    "google-adk>=1.18.0",
    "litellm>=1.80.0",
    "mcp[cli]>=1.21.0",
    "mysql-connector-python>=9.5.0",
    "openai>=2.7.2",
    "python-dotenv>=1.2.1",
]
